added new component to make some api calls to populate child components and routes w/ burnettk
